requirements.txt
boto3
botocore
strands-agents
bedrock-agentcore
bedrock-agentcore-starter-toolkit
python-dotenv
requests